+++ libswirl/reios/reios_elf.cpp
@@ -52,7 +52,7 @@ bool reios_loadElf(const string& elf) {
 		len = elf_getProgramHeaderFileSize(elfFile, i);
 		src = (uint64_t)(uintptr_t)elfFile + elf_getProgramHeaderOffset(elfFile, i);
 		
-		u8* ptr = GetMemPtr(dest, len);
+		u8* ptr = GetMemPtr((u32)dest, (u32)len);
 
 		memcpy((void*)ptr, (void*)(uintptr_t)src, len);
 		ptr += len;
