+++ libswirl/rend/gles/glesrend.cpp
@@ -1358,7 +1358,7 @@ u8* loadPNGData(const string& fname, int &width, int &height)
 	png_read_update_info(png_ptr, info_ptr);
 
 	// Row size in bytes.
-	int rowbytes = png_get_rowbytes(png_ptr, info_ptr);
+	int rowbytes = (int)png_get_rowbytes(png_ptr, info_ptr);
 
 	// Allocate the image_data as a big block, to be given to opengl
 	png_byte *image_data = new png_byte[rowbytes * height];
